refactor: remove src directory
